+++ core/src/main/java/com/graphhopper/util/GHUtility.java
@@ -206,9 +206,9 @@ public class GHUtility {
     /**
      * @param speed if null a random speed will be assigned to every edge
      */
-    public static void buildRandomGraph(Graph graph, Random random, int numNodes, double meanDegree, boolean allowLoops,
+    public static void buildRandomGraph(Graph graph, Random random, int numNodes, double meanDegree,
                                         boolean allowZeroDistance, BooleanEncodedValue accessEnc, DecimalEncodedValue speedEnc, Double speed,
-                                        double pNonZeroLoop, double pBothDir, double pRandomDistanceOffset) {
+                                        double pBothDir, double pRandomDistanceOffset) {
         if (numNodes < 2 || meanDegree < 1) {
             throw new IllegalArgumentException("numNodes must be >= 2, meanDegree >= 1");
         }
@@ -224,14 +224,9 @@ public class GHUtility {
         while (numEdges < totalNumEdges) {
             int from = random.nextInt(numNodes);
             int to = random.nextInt(numNodes);
-            if (!allowLoops && from == to) {
+            if (from == to)
                 continue;
-            }
             double distance = GHUtility.getDistance(from, to, graph.getNodeAccess());
-            // allow loops with non-zero distance
-            if (from == to && random.nextDouble() < pNonZeroLoop) {
-                distance = random.nextDouble() * 1000;
-            }
             if (!allowZeroDistance) {
                 distance = Math.max(0.001, distance);
             }
@@ -497,14 +492,14 @@ public class GHUtility {
     /**
      * Creates an edge key, i.e. an integer number that encodes an edge ID and the direction of an edge
      */
-    public static int createEdgeKey(int edgeId, boolean isLoop, boolean reverse) {
+    public static int createEdgeKey(int edgeId, boolean reverse) {
         // edge state in storage direction -> edge key is even
         // edge state against storage direction -> edge key is odd
-        return (edgeId << 1) + ((reverse && !isLoop) ? 1 : 0);
+        return (edgeId << 1) + (reverse ? 1 : 0);
     }
 
     /**
-     * Returns the edgeKey of the opposite direction, be careful not to use this for loops!
+     * Returns the edgeKey of the opposite direction
      */
     public static int reverseEdgeKey(int edgeKey) {
         return edgeKey % 2 == 0 ? edgeKey + 1 : edgeKey - 1;
@@ -562,12 +557,8 @@ public class GHUtility {
     }
 
     public static double calcWeightWithTurnWeightWithAccess(Weighting weighting, EdgeIteratorState edgeState, boolean reverse, int prevOrNextEdgeId) {
-        if (edgeState.getBaseNode() == edgeState.getAdjNode()) {
-            if (weighting.edgeHasNoAccess(edgeState, false) && weighting.edgeHasNoAccess(edgeState, true))
-                return Double.POSITIVE_INFINITY;
-        } else if (weighting.edgeHasNoAccess(edgeState, reverse)) {
+        if (weighting.edgeHasNoAccess(edgeState, reverse))
             return Double.POSITIVE_INFINITY;
-        }
         return calcWeightWithTurnWeight(weighting, edgeState, reverse, prevOrNextEdgeId);
     }
 
@@ -710,16 +701,11 @@ public class GHUtility {
         IntIndexedContainer refNodes = refPath.calcNodes();
         IntIndexedContainer pathNodes = path.calcNodes();
         if (!refNodes.equals(pathNodes)) {
-            // sometimes paths are only different because of a zero weight loop. we do not consider these as strict
-            // violations, see: #1864
-            boolean isStrictViolation = !ArrayUtil.withoutConsecutiveDuplicates(refNodes).equals(ArrayUtil.withoutConsecutiveDuplicates(pathNodes));
             // sometimes there are paths including an edge a-c that has the same distance as the two edges a-b-c. in this
             // case both options are valid best paths. we only check for this most simple and frequent case here...
             if (path.getGraph() != refPath.getGraph())
                 fail("path and refPath graphs are different");
-            if (pathsEqualExceptOneEdge(path.getGraph(), refNodes, pathNodes))
-                isStrictViolation = false;
-            if (isStrictViolation)
+            if (!pathsEqualExceptOneEdge(path.getGraph(), refNodes, pathNodes))
                 strictViolations.add("wrong nodes " + source + "->" + target + "\nexpected: " + refNodes + "\ngiven:    " + pathNodes);
         }
         return strictViolations;